Conflicts:
	sicf-consultas-generales/src/main/java/ob/unibanca/sicf/consultasgenerales/model/swdmplog/TxnSwdmplog.java
	sicf-consultas-generales/src/main/java/ob/unibanca/sicf/consultasgenerales/model/swdmplog/TxnSwdmplogDetalle.java

+++ sicf-consultas-generales/src/main/java/ob/unibanca/sicf/consultasgenerales/model/incomingvisa/IncomingVisaTC10TCR0.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @EqualsAndHashCode(callSuper = false)
 public class IncomingVisaTC10TCR0 {
-	Integer secuenciaIncoming;
+	Long secuenciaIncoming;
 	Date fechaProceso;
 	String transactionCode;
 	String descTransactionCode;
