+++ sicf-consultas-generales/src/main/java/ob/unibanca/sicf/consultasgenerales/model/incomingvisa/IncomingVisaTC5TCR1.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class IncomingVisaTC5TCR1 {
-	private Integer idSecuenciaIncoming;
+	private Long idSecuenciaIncoming;
 	private String transactionCode;
 	private String descTransactionCode;
 	private String transCodeQualifier;
